Implement basic dynamic sampling
The dynsampler code is somewhat unnecessary right now, but
there are a bunch of other strategies in there which I expect
we'll want to explore - so coding to that interface right now
seems reasonable.

Note that this approach doesn't actually sample child spans
at the same rate as traces with non-default rates. I asked in
the honeycomb slack, and to do that we apparently need to do
some sort of buffering in memory or see if we can get access
to their server-side sampling beta