Attach OpenTelemetry Baggage to traces when available
Spec:
https://github.com/open-telemetry/opentelemetry-specification/blob/14b5b6a944e390e368dd2e2ef234d220d8287d19/specification/baggage/api.md

Baggage is a map of values used for telemetry purposes. When
propogated through an HTTP header, deserialize the Baggage and add its
fields to the trace.

For the job submission API, persist the baggage to the job table so
that it can be used to annotate the trace for the task claiming API.