Adds more sensible query timeout in dispatch.
Whilst testing the timeout it became clear that the multierror used in the
dispatch fan out to collect the errors would itself trace as error even if all
the errors in its list would not. The error is change to be a custom type
that will inspect its contained errors for whether it should be traced as an
error or not.