Add GaugeProducer interface
This interface is intended to work similarly to the MetricProducer
interface, but with Gauges. It is designed to enable sending multiple
gauges of the same name, but with different tags and values.

The MetricProducer interface has been left in place for
backwards-compatibility, but this should be considered the new way to
emit gauge metrics on a regular interval.