Add standard metrics for warning and error
This insects a span for the warning and error fields and produces a count metric
tagged with type:o11y.

We use field names with an _error suffix to indicate a problem span but only in
a path that can still make progress, e.g. a fail open path.

Since the field names are varied it is hard to add monitors, so this detects those
fields, adds a standard 'failure' field to the span with the prefix part as the value
and also produces a failure metric and uses the prefix part as a tag.