[EXEC-3120] Allow non-mandatory messages to be published
This should allow messages to be posted even though no exchanges exist.
Optional messages allow us to start emitting messages even though nobody
listens for them yet, which allows people to start working with the new
messages before we make them mandatory.

Definitely the "mandatory" should happen afterwards, but for now
it allows us to be less dependant on when a team is ready to start
listening for messages before we deliver those.